+++ module/Application/src/Application/Controller/NewsController.php
@@ -50,21 +50,21 @@ final class NewsController extends IndexController
      */
     protected function postAction()
     {
-        $this->view->setTemplate("application/news/post");
+        $this->getView()->setTemplate("application/news/post");
         $escaper = new \Zend\Escaper\Escaper('utf-8');
         $post = (string) $escaper->escapeUrl($this->getParam("post"));
         $content = $this->getTable("content");
 
         if (!empty($post)) {
             $new = $content->fetchList(false, ["title", "text", "date", "preview"], ["type" => 1, "menu" => 0, "titleLink" => $post, "language" => $this->language()], "AND", null, "date DESC")->getDataSource()->current();
-            $this->view->new = $new;
+            $this->getView()->new = $new;
             $this->initMetaTags($new);
         } else {
             $news = $content->fetchList(true, ["title", "titleLink", "text", "date", "preview"], ["type" => 1, "menu" => 0, "language" => $this->language()], "AND", null, "date DESC");
             $news->setCurrentPageNumber((int)$this->getParam('page', 1));
             $news->setItemCountPerPage(2); // must be set from db
-            $this->view->news = $news;
+            $this->getView()->news = $news;
         }
-        return $this->view;
+        return $this->getView();
     }
 }